+++ clang/lib/Format/TokenAnnotator.cpp
@@ -3433,6 +3433,7 @@ bool TokenAnnotator::mustBreakForReturnType(const AnnotatedLine &Line) const {
   switch (Style.AlwaysBreakAfterReturnType) {
   case FormatStyle::RTBS_None:
   case FormatStyle::RTBS_AllowShortType:
+  case FormatStyle::RTBS_ExceptShortType:
     return false;
   case FormatStyle::RTBS_All:
   case FormatStyle::RTBS_TopLevel: